Move gems_runner to runner.py
